pyproject.toml
[project]
name = "vision-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.1",
    "h5py>=3.14.0",
    "imageio>=2.37.0",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tensorboard>=2.19.0",
    "tensorboard-plugin-profile>=2.20.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
